[CL-2362] Move components into the containers
