+++ front/app/modules/commercial/custom_topics/admin/containers/TopicsSettings/all/TopicTermConfig.tsx
